requirements.txt
docker
pyyaml
argparse
halo
